refactor(data-import): inline types and match-headers into hook
Drop standalone types.ts and match-headers.ts in data-import/. Schemas
collapse into use-data-import.tsx with Zod for data shapes
(rawImportDataSchema, importableColumnSchema) and TS interfaces for
configs carrying callbacks. autoMatch + getImportableColumns become
private to the hook (single caller).
